Changed startCapture function to take an int instead of a long for number of frames.  If we need more than 2 billion frames taken, we should probably just do continuous capture.

+++ core/trunk/src/edu/cmu/ri/crw/VehicleServer.java
@@ -32,7 +32,7 @@ public interface VehicleServer {
         
         public void addCameraListener(CameraListener l);
 	public void removeCameraListener(CameraListener l);
-	public void startCamera(long numFrames, double interval, int width, int height);
+	public void startCamera(int numFrames, double interval, int width, int height);
 	public void stopCamera();
 	public CameraState getCameraStatus();
